Ref: Move all snapshot related functions to a new file
